requirements.txt
annotated-types==0.6.0
backcall==0.1.0
beautifulsoup4==4.12.3
bitarray==2.7.2
blis==0.7.11
boto3==1.9.239
botocore==1.12.239
cachetools==5.3.3
catalogue==2.0.10
certifi==2019.9.11
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
cymem==2.0.8
decorator==4.4.1
docutils==0.15.2
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1-py3-none-any.whl#sha256=ab70aeb6172cde82508f7739f35ebc9918a3d07debeed637403c8f794ba3d3dc
google-ai-generativelanguage==0.4.0
google-api-core==2.18.0
google-api-python-client==2.123.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.4.1
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
httplib2==0.22.0
idna==2.8
ipdb==0.12.3
ipython==7.10.2
ipython-genutils==0.2.0
jedi==0.15.1
Jinja2==3.1.3
jmespath==0.9.4
langcodes==3.3.0
lxml==5.1.0
MarkupSafe==2.1.5
murmurhash==1.0.10
numpy==1.20.0
packaging==24.0
parso==0.5.2
pexpect==4.7.0
pickleshare==0.7.5
Pillow==9.3.0
pkg_resources==0.0.0
preshed==3.0.9
prompt-toolkit==3.0.2
proto-plus==1.23.0
protobuf==4.25.3
ptyprocess==0.6.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.5.2
pyparsing==3.1.2
python-dateutil==2.8.0
pytorch-pretrained-bert==0.6.2
regex==2022.10.31
requests==2.22.0
rsa==4.9
s3transfer==0.2.1
scipy==1.6.0
six==1.16.0
smart-open==6.4.0
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
thinc==8.2.3
torch==1.12.0
torchvision==0.13.0
tqdm==4.66.2
traitlets==4.3.3
typer==0.9.0
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==1.25.6
wasabi==1.1.2
wcwidth==0.1.7
weasel==0.3.4
